+++ CHANGELOG.md
@@ -21,6 +21,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog-emoji/master/CHANGELOG.md
 -->
 
+## [Unreleased]
+
+### 📚 Documentation
+
+- Add ShellUI brand favicon (ICO + PNG sizes) to the Docusaurus docs site.
+
 ## [0.4.1] - 2026-08-18
 
 ### 🐛 Bug Fixes